moved homepage components into componenets folder
