+++ src/components/pokedex-display/pokemon-card/style.ts
